# docs/swagger.yaml
definitions:
...
  helper.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  helper.Response:
    properties:
      data: {}
      message:
        type: string
    type: object
...
  service.CreateCategoryInput:
    properties:
      department_id:
        type: integer
      description:
        type: string
      name:
        type: string
    type: object
  service.CreateDistrictInput:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  service.LoginInput:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
...
  service.RegisterInput:
    properties:
      email:
        type: string
      home_district_id:
        type: integer
      name:
        maxLength: 100
        minLength: 3
        type: string
      nik:
        type: string
      password:
        minLength: 8
        type: string
      phone_number:
        maxLength: 15
        minLength: 9
        type: string
    required:
    - email
    - name
    - nik
    - password
    - phone_number
    type: object
  service.ToggleCategoryStatusInput:
    properties:
      is_active:
        type: boolean
    type: object
...
  service.UpdateCategoryInput:
    properties:
      department_id:
        type: integer
      description:
        type: string
      name:
        type: string
    type: object
  service.UpdateDistrictInput:
    properties:
      is_active:
        type: boolean
    type: object
...
info:
  contact: {}
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login user
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: Login
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user account
      parameters:
      - description: Register data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.RegisterInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: Register user
      tags:
      - Auth
  /categories:
    get:
      description: Get all categories
      parameters:
      - description: Show only active categories
        in: query
        name: active_only
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: Get all categories
      tags:
      - Category
    post:
      consumes:
      - application/json
      description: Create a new category
      parameters:
      - description: Category data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.CreateCategoryInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create category
      tags:
      - Category
  /categories/{id}:
    get:
      description: Retrieve category detail
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: Get category by ID
      tags:
      - Category
    put:
      consumes:
      - application/json
      description: Update category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Category data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UpdateCategoryInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update category
      tags:
      - Category
  /categories/{id}/status:
    patch:
      consumes:
      - application/json
      description: Enable or disable category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Status data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.ToggleCategoryStatusInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Toggle category status
      tags:
      - Category
...
  /districts:
    get:
      description: Retrieve all districts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: Get all districts
      tags:
      - District
    post:
      consumes:
      - application/json
      description: Create a new district
      parameters:
      - description: District data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.CreateDistrictInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create district
      tags:
      - District
  /districts/{id}:
    get:
      description: Retrieve district detail
      parameters:
      - description: District ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: Get district by ID
      tags:
      - District
    put:
      consumes:
      - application/json
      description: Update district
      parameters:
      - description: District ID
        in: path
        name: id
        required: true
        type: integer
      - description: District data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UpdateDistrictInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update district
      tags:
      - District
...
swagger: "2.0"