# internal/controller/auth_controller.go
package controller

import (
	"github.com/labstack/echo/v4"

	"github.com/manciniraka/urbioxe/internal/helper"
	"github.com/manciniraka/urbioxe/internal/service"
)

type AuthController struct {
	authService service.AuthService
}

func NewAuthController(
	authService service.AuthService,
) *AuthController {
	return &AuthController{
		authService: authService,
	}
}

type LoginResponse struct {
	Token string `json:"token"`
}

// Register godoc
//
//	@Summary		Register user
//	@Description	Register a new user account
//	@Tags			Auth
//	@Accept			json
//	@Produce		json
//	@Param			request	body		service.RegisterInput	true	"Register data"
//	@Success		201		{object}	helper.Response
//	@Failure		400		{object}	helper.ErrorResponse
//	@Failure		500		{object}	helper.ErrorResponse
//	@Router			/auth/register [post]
func (ac *AuthController) Register(c echo.Context) error {
	var input service.RegisterInput

	if err := c.Bind(&input); err != nil {
		return helper.BadRequest(
			c,
			"invalid request body",
		)
	}

	if err := c.Validate(&input); err != nil {
		return helper.BadRequest(
			c,
			err.Error(),
		)
	}

	user, err := ac.authService.Register(input)
	if err != nil {
		return helper.HandleError(
			c,
			err,
		)
	}

	return helper.Created(
		c,
		"user registered successfully",
		user,
	)
}

// Login godoc
//
//	@Summary		Login
//	@Description	Login user
//	@Tags			Auth
//	@Accept			json
//	@Produce		json
//	@Param			request	body		service.LoginInput	true	"Login credentials"
//	@Success		200		{object}	helper.Response
//	@Failure		400		{object}	helper.ErrorResponse
//	@Failure		401		{object}	helper.ErrorResponse
//	@Failure		500		{object}	helper.ErrorResponse
//	@Router			/auth/login [post]
func (ac *AuthController) Login(c echo.Context) error {
	var input service.LoginInput

	if err := c.Bind(&input); err != nil {
		return helper.BadRequest(
			c,
			"invalid request body",
		)
	}

	if err := c.Validate(&input); err != nil {
		return helper.BadRequest(
			c,
			err.Error(),
		)
	}

	token, err := ac.authService.Login(input)
	if err != nil {
		return helper.HandleError(
			c,
			err,
		)
	}

	return helper.Success(
		c,
		"login successfully",
		LoginResponse{
			Token: token,
		},
	)
}